update envs layers
