+++ ai/train/trainer.py
@@ -1,7 +1,7 @@
 """Training entry point.
 
-``Dispatches to MJX GPU-parallel RL, imitation learning, or evaluation
-based on the TrainingConfig in the unified Config.
+Dispatches to the appropriate simulator backend (MJX, Isaac Sim, ...)
+and training method (RL, imitation, eval) based on the unified Config.
 
 Usage:
     bazel run //ai/train:trainer -- \
@@ -31,6 +31,57 @@ def _load_training_config(path: str) -> training_pb2.TrainingConfig:
     return full.ai.training
 
 
+def _run_rl(config: training_pb2.TrainingConfig) -> None:
+    """Dispatch RL training to the right simulator backend."""
+    if config.environment != training_pb2.TRAINING_ENV_SIMULATION:
+        glog.warning("RL training currently only supports simulation environment")
+
+    backend = config.simulator_backend
+    model_path = config.model_path
+
+    if backend == training_pb2.SIM_BACKEND_MJX or backend == training_pb2.SIM_BACKEND_INVALID:
+        if backend == training_pb2.SIM_BACKEND_INVALID:
+            glog.warning(
+                "simulator_backend not set, defaulting to MJX. "
+                "Set simulator_backend: SIM_BACKEND_MJX in your config."
+            )
+        from ai.train.mjx_rl import run as run_mjx
+        run_mjx(config.rl, model_path)
+
+    elif backend == training_pb2.SIM_BACKEND_ISAAC_SIM:
+        raise NotImplementedError(
+            "Isaac Sim backend is not yet implemented. "
+            "Set simulator_backend: SIM_BACKEND_MJX to use MJX."
+        )
+
+    else:
+        raise ValueError(
+            f"Unknown simulator_backend: {backend}. "
+            f"Use SIM_BACKEND_MJX or SIM_BACKEND_ISAAC_SIM."
+        )
+
+
+def _run_eval(config: training_pb2.TrainingConfig) -> None:
+    """Dispatch evaluation to the right simulator backend."""
+    if config.environment != training_pb2.TRAINING_ENV_SIMULATION:
+        raise NotImplementedError("Real-world evaluation is not yet implemented")
+
+    backend = config.simulator_backend
+    model_path = config.model_path
+
+    if backend == training_pb2.SIM_BACKEND_MJX or backend == training_pb2.SIM_BACKEND_INVALID:
+        from ai.train.mjx_rl import eval_mjx
+        eval_mjx(config.eval, model_path)
+
+    elif backend == training_pb2.SIM_BACKEND_ISAAC_SIM:
+        raise NotImplementedError(
+            "Isaac Sim evaluation is not yet implemented."
+        )
+
+    else:
+        raise ValueError(f"Unknown simulator_backend: {backend}")
+
+
 def main(argv):
     try:
         argv = FLAGS(argv)
@@ -46,18 +97,13 @@ def main(argv):
 
     env_name = training_pb2.TrainingEnvironment.Name(config.environment)
     method_name = training_pb2.TrainingMethod.Name(config.method)
-    glog.info(f"Training config: environment={env_name}, method={method_name}")
+    backend_name = training_pb2.SimulatorBackend.Name(config.simulator_backend)
+    glog.info(f"Training config: environment={env_name}, method={method_name}, "
+              f"backend={backend_name}")
 
     if config.method == training_pb2.TRAINING_METHOD_RL:
-        # TODO(hmoon): Implement real-world RL training
-        if config.environment != training_pb2.TRAINING_ENV_SIMULATION:
-            glog.warning("RL training currently only supports simulation environment")
+        _run_rl(config)
 
-        # TODO(hmoon): Add Issac Sim support here
-        from ai.train.mjx_rl import run as run_mjx
-        run_mjx(config.rl)
-
-    # TODO(hmoon): Implement imitation learning
     elif config.method == training_pb2.TRAINING_METHOD_IMITATION:
         glog.info("Imitation learning — not yet implemented")
         glog.info(f"  dataset_path: {config.imitation.dataset_path}")
@@ -65,20 +111,7 @@ def main(argv):
         raise NotImplementedError("Imitation learning is not yet implemented")
 
     elif config.method == training_pb2.TRAINING_METHOD_EVAL:
-        if config.environment != training_pb2.TRAINING_ENV_SIMULATION:
-            raise NotImplementedError("Real-world evaluation is not yet implemented")
-
-        from ai.train.mjx_rl import eval_mjx
-
-        rl_config = training_pb2.RLConfig(
-            task=config.eval.task,
-            checkpoint_path=config.eval.checkpoint_path,
-            num_eval_episodes=config.eval.num_episodes,
-            total_timesteps=0,
-        )
-        if config.eval.render:
-            rl_config.render = True
-        eval_mjx(rl_config)
+        _run_eval(config)
 
     else:
         raise ValueError(f"Unknown training method: {config.method}")
