+++ config/config_preset/ant_train_mjx.pbtxt
@@ -6,6 +6,8 @@ ai {
   training {
     environment: TRAINING_ENV_SIMULATION
     method: TRAINING_METHOD_RL
+    model_path: "simulation/models/ant_mjx.xml"
+    simulator_backend: SIM_BACKEND_MJX
     rl {
       task: "ant"
       frame_skip: 5
