+++ config/config_preset/so_arm100_eval.pbtxt
@@ -7,6 +7,8 @@ ai {
   training {
     environment: TRAINING_ENV_SIMULATION
     method: TRAINING_METHOD_EVAL
+    model_path: "simulation/models/so_arm100_pick_place_mjx.xml"
+    simulator_backend: SIM_BACKEND_MJX
     eval {
       task: "pick_place"
       checkpoint_path: "pick_place_mjx_ppo.msgpack"
